% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EigenVectors}
\alias{EigenVectors}
\title{Eigenvectors Extraction}
\usage{
EigenVectors(X, D)
}
\arguments{
\item{X}{covariance matrix to exttrat first D unitlength eigenvectors}

\item{D}{number of orthonormal eigenvectors to return}
}
\value{
first D eigenvectors of the covariance matrix
}
\description{
Returns the first D real eigenvectors of a given covariance matrix
}
\examples{
>> X_i_jk = generate_dataset(random_state=0)$X_i_jk
>> norm(EigenVectors(X_i_jk \%*\% t(X_i_jk),3),'F')  # 1.732051
}
