% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous.R
\name{fit.3fkmeans}
\alias{fit.3fkmeans}
\alias{fit.3fkmeans,simultaneous-method}
\title{3FKMeans Model}
\usage{
fit.3fkmeans(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

\S4method{fit.3fkmeans}{simultaneous}(model)
}
\arguments{
\item{model}{Initialized simultaneous model.}

\item{X_i_jk}{Matricized tensor along mode-1 (I objects).}

\item{full_tensor_shape}{Dimensions of the tensor in full-space.}

\item{reduced_tensor_shape}{Dimensions of tensor in the reduced-space.}
}
\description{
Implements simultaneous version of TWFCTA
}
\examples{
X_i_jk = generate_dataset()$X_i_jk
model = simultaneous()
tfkmeans = fit.3fkmeans(model, X_i_jk, c(8,5,4), c(3,3,2))

}
\references{
\insertRef{tucker1966}{simuclustfactor}
\insertRef{VichiRocciKiers}{simuclustfactor}
}
