% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous.R
\name{fit.ct3clus}
\alias{fit.ct3clus}
\alias{fit.ct3clus,simultaneous-method}
\title{CT3Clus Model}
\usage{
fit.ct3clus(
  model,
  X_i_jk,
  full_tensor_shape,
  reduced_tensor_shape,
  alpha = 0.5
)

\S4method{fit.ct3clus}{simultaneous}(
  model,
  X_i_jk,
  full_tensor_shape,
  reduced_tensor_shape,
  alpha = 0.5
)
}
\arguments{
\item{model}{Initialized simultaneous model.}

\item{X_i_jk}{Matricized tensor along mode-1 (I objects).}

\item{full_tensor_shape}{Dimensions of the tensor in full space.}

\item{reduced_tensor_shape}{Dimensions of tensor in the reduced space.}

\item{alpha}{0<alpha>1 hyper parameter. Model is T3Clus when alpha=1 and 3FKMeans when alpha=0.}
}
\description{
Implements simultaneous T3Clus and 3FKMeans integrating
an alpha value
}
\examples{
X_i_jk = generate_dataset()$X_i_jk
model = simultaneous()
ct3clus = fit.ct3clus(model, X_i_jk, c(8,5,4), c(3,3,2), alpha=0.5)

}
\references{
\insertRef{tucker1966}{simuclustfactor}
\insertRef{t3clus}{simuclustfactor}
\insertRef{3FKMeans}{simuclustfactor}
\insertRef{VichiRocciKiers}{simuclustfactor}
}
