% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.R
\name{fit.twcfta}
\alias{fit.twcfta}
\alias{fit.twcfta,tandem-method}
\title{TWCFTA model}
\usage{
fit.twcfta(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

\S4method{fit.twcfta}{tandem}(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)
}
\arguments{
\item{model}{Initialized tandem model.}

\item{X_i_jk}{Matricized tensor along mode-1 (I objects).}

\item{full_tensor_shape}{Dimensions of the tensor in full space.}

\item{reduced_tensor_shape}{Dimensions of tensor in the reduced space.}
}
\description{
Implements K-means clustering and afterwards factorial reduction
in a sequential fashion.
}
\examples{
X_i_jk = generate_dataset()$X_i_jk
model = tandem()
twcfta = fit.twcfta(model, X_i_jk, c(8,5,4), c(3,3,2))

}
\references{
\insertRef{tandemModels}{simuclustfactor}
\insertRef{tucker1966}{simuclustfactor}
}
