% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.R
\name{fit.twfcta}
\alias{fit.twfcta}
\alias{fit.twfcta,tandem-method}
\title{TWFCTA model}
\usage{
fit.twfcta(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

\S4method{fit.twfcta}{tandem}(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)
}
\arguments{
\item{model}{Initialized tandem model.}

\item{X_i_jk}{Matricized tensor along mode-1 (I objects).}

\item{full_tensor_shape}{Dimensions of the tensor in full space.}

\item{reduced_tensor_shape}{Dimensions of tensor in the reduced space.}
}
\description{
Implements factorial reduction and then K-means clustering
in a sequential fashion.
}
\examples{
X_i_jk = generate_dataset()$X_i_jk
model = tandem()
twfCta = fit.twfcta(model, X_i_jk, c(8,5,4), c(3,3,2))

}
\references{
\insertRef{tandemModels}{simuclustfactor}
\insertRef{tucker1966}{simuclustfactor}
}
