% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{Fold}
\alias{Fold}
\title{Folding matrix back to tensor by mode.}
\usage{
Fold(X, mode, shape)
}
\arguments{
\item{X}{data matrix to fold}

\item{mode}{mode of operation}

\item{shape}{dimension of original tensor}
}
\value{
tensor
}
\description{
X_i_jk => X_i_j_k, X_j_ki => X_i_j_k, X_k_ij => X_i_j_k
}
\examples{
>> X_i_jk = generate_dataset()$X_i_jk
>> Fold(X_i_jk, mode=1, shape=c(I=8,J=5,K=4)) # X_i_j_k
>> Fold(X_j_ki, mode=2, shape=c(I=8,J=5,K=4)) # X_i_j_k
>> Fold(X_k_ij, mode=3, shape=c(I=8,J=5,K=4)) # X_i_j_k
}
