% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Dataset Generation for simulation}
\usage{
generate_dataset(
  I = 8,
  J = 5,
  K = 4,
  G = 3,
  Q = 3,
  R = 2,
  mean = 0,
  stdev = 0.5,
  random_state = NULL
)
}
\arguments{
\item{I}{number of objects.}

\item{J}{number of variables per occasion}

\item{K}{number of occasions}

\item{G}{number of clusters}

\item{Q}{number of factors for the variables}

\item{R}{number of factors for the occasions}

\item{mean}{noise effect mean}

\item{stdev}{noise effect level/spread/standard deviation}

\item{random_state}{seed for random number generation}
}
\value{
Z_i_jk: component scores in the full space

X_i_jk: dataset with noise level set to stdev specified

Y_g_qr: centroids matrix in the reduced space

U_i_g: stochastic membership function matrix

B_j_q: objects component scores matrix

C_k_r: occasions component scores matrix
}
\description{
Generate G clustered synthetic dataset of I objects measured on J variables
for K occasions with additive noise.
}
\examples{
>> X_i_jk = generate_dataset(random_state=0)
}
