% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_rmfm}
\alias{generate_rmfm}
\title{Random Membership Function Matrix Generator}
\usage{
generate_rmfm(I, G, seed = NULL)
}
\arguments{
\item{I}{Number of objects.}

\item{G}{Number of groups/clusters.}

\item{seed}{Seed for random number generation.}
}
\value{
U_i_g, binary stochastic membership matrix.
}
\description{
Generates random binary stochastic membership function matrix for the I objects.
}
\examples{
generate_rmfm(I=8,G=3)

}
