% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudof.reduced}
\alias{pseudof.reduced}
\title{PseudoF Score in the Reduced-Space}
\usage{
pseudof.reduced(bss, wss, full_tensor_shape, reduced_tensor_shape)
}
\arguments{
\item{bss}{Between sums of squared deviations between clusters.}

\item{wss}{Within sums of squared deviations within clusters.}

\item{full_tensor_shape}{Dimensions of the tensor in the original space.}

\item{reduced_tensor_shape}{Dimension of the tensor in the reduced space.}
}
\value{
PseudoF score
}
\description{
Computes the PseudoF score in the reduced space.
}
\examples{
pseudof.reduced(12,6,c(8,5,4),c(3,3,2))

}
\references{
\insertRef{pseudoF}{simuclustfactor}
}
