% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PseudoF_Full}
\alias{PseudoF_Full}
\title{PseudoF sore in Full space}
\usage{
PseudoF_Full(bss, wss, full_tensor_shape, reduced_tensor_shape)
}
\arguments{
\item{bss}{between sums of squared deviations between clusters}

\item{wss}{within sums of squared deviations within clusters}

\item{full_tensor_shape}{dimensions of the tensor in the original space}

\item{reduced_tensor_shape}{dimension of the tensor in the reduced space.}
}
\value{
PseudoF score
}
\description{
computes the PseudoF score in the full space.
}
\examples{
PseudoF_full(12,6,c(8,5,4),c(3,3,2))
}
