% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RandomMembershipMatrix}
\alias{RandomMembershipMatrix}
\title{Random Membership Matrix Generator}
\usage{
RandomMembershipMatrix(I, G, seed = NULL)
}
\arguments{
\item{I}{number of objects.}

\item{G}{number of groups/clusters.}

\item{seed}{random_state seed.}
}
\value{
U_i_g, binary stochastic membership matrix
}
\description{
Generates binary stochastic membership function matrix for the I objects.
}
\examples{
>> RandomMembershipMatrix(I=8,G=3)
}
