% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.R
\docType{class}
\name{tandem-class}
\alias{tandem-class}
\title{Tandem Class}
\description{
Tandem Class
}
\section{Slots}{

\describe{
\item{\code{seed}}{Seed for random sequence generation. Defaults to None.}

\item{\code{verbose}}{logical. Whether to display executions output or not. Defaults to False.}

\item{\code{init}}{character. The parameter initialization method. Defaults to 'svd'.}

\item{\code{n_max_iter}}{numeric. Maximum number of iterations. Defaults to 10.}

\item{\code{n_loops}}{numeric. Number of initialization to guarantee global results. Defaults to 10.}

\item{\code{tol}}{numeric. Tolerance level/acceptable error. Defaults to 1e-5.}

\item{\code{U_i_g}}{matrix. (I,G) initial stochastic membership function matrix.}

\item{\code{B_j_q}}{matrix. (J,Q) initial component weight matrix for variables.}

\item{\code{C_k_r}}{matrix. (K,R) initial component weight matrix for occasions.}
}}

