% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.R
\name{fit.twcfta,tandem-method}
\alias{fit.twcfta,tandem-method}
\alias{TWCFTA,twcfta-model}
\alias{fit.twfcta,tandem-method}
\alias{TWFCTA,twfcta-model}
\title{TWCFTA model}
\usage{
\S4method{fit.twcfta}{tandem}(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

\S4method{fit.twfcta}{tandem}(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)
}
\description{
TWCFTA model

TWFCTA model
}
