% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.R
\docType{methods}
\name{fit.twfcta}
\alias{fit.twfcta}
\title{TWFCTA model}
\usage{
fit.twfcta(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)
}
\arguments{
\item{model}{initialized tandem model}

\item{X_i_jk}{matricized tensor along mode-1 (I objects)}

\item{full_tensor_shape}{dimensions of the tensor in full space}

\item{reduced_tensor_shape}{dimensions of tensor in the reduced space}
}
\description{
Implements factorial reduction and then K-means clustering
in a sequential fashion
}
\examples{
X_i_jk = generate_dataset()
model = tandem()
twfCta = fit.twfcta(model, X_i_jk, c(8,5,4), c(3,3,2))

}
