% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.R
\name{tandem}
\alias{tandem}
\title{Initialization for the tandem models.}
\usage{
tandem(
  random_state = NULL,
  verbose = TRUE,
  init = "svd",
  n_max_iter = 10,
  n_loops = 10,
  tol = 1e-05,
  U_i_g = NULL,
  B_j_q = NULL,
  C_k_r = NULL
)
}
\arguments{
\item{random_state}{for random number generation}

\item{verbose}{flag to display iteration outputs}

\item{init}{parameter initialization method, 'svd' or 'random'}

\item{n_max_iter}{maximum number of iteration to optimize the objective function}

\item{n_loops}{maximum number of loops/runs for global results}

\item{tol}{allowable tolerance to check convergence}

\item{U_i_g}{initial membership function matrix for the objects}

\item{B_j_q}{initial component scores matrix for the variables}

\item{C_k_r}{initial component sores matrix for the occasions}
}
\description{
Initialization for the tandem models.
}
