% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{Unfold}
\alias{Unfold}
\title{Tensor matricization}
\usage{
Unfold(tensor, mode)
}
\arguments{
\item{tensor}{three-mode tensor array}

\item{mode}{mode of operation}
}
\value{
matrix
}
\description{
Unfold/matricize tensor. convert matrix to tensor by mode.
}
\examples{
>> X_i_jk = generate_dataset()$X_i_jk
>> X_i_j_k = Fold(X_i_jk, mode=1, shape=c(I=8,J=5,K=4))
>> Unfold(X=X_i_j_k, mode=1) # X_i_jk
>> Unfold(X=X_i_j_k, mode=2) # X_j_ki
>> Unfold(X=X_i_j_k, mode=3) # X_k_ij
}
