% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{fold}
\alias{fold}
\title{Folding Matrix to Tensor by Mode.}
\usage{
fold(X, mode, shape)
}
\arguments{
\item{X}{Data matrix to fold.}

\item{mode}{Mode of operation.}

\item{shape}{Dimension of original tensor.}
}
\value{
X_i_j_k Three-mode tensor.
}
\description{
X_i_jk => X_i_j_k, X_j_ki => X_i_j_k, X_k_ij => X_i_j_k
}
\examples{
X_i_jk = generate_dataset()$X_i_jk
X_i_j_k = fold(X_i_jk, mode=1, shape=c(I=8,J=5,K=4)) # X_i_j_k

}
