% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OneKMeans}
\alias{OneKMeans}
\title{One-run of the K-means clustering technique}
\usage{
OneKMeans(Y_i_qr, G, U_i_g = NULL, seed = NULL)
}
\arguments{
\item{Y_i_qr}{# input data to group/cluster.}

\item{G}{# number of clusters to find.}

\item{U_i_g}{current membership function matrix with empty cluster}

\item{seed}{# random_state for random values generation.}

\item{LC}{largest cluster index}

\item{EC}{empty cluster index to share members of LC with}
}
\value{
updated membership matrix U_i_g

U_i_g, the updated membership matrix.
}
\description{
initializes centroids based on input centroid or randomly. iterate once over
input data to update the centroids assigning objects to the closest centroids.
}
\examples{
>> X_i_jk = generate_dataset(random_state=0)$X_i_jk
>> OneKMeans(X_i_jk, G=5)
Split largest cluster with chosen empty cluster.

if there is an empty cluster share members of largest cluster with
empty cluster.

}
