% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{unfold}
\alias{unfold}
\title{Tensor Matricization}
\usage{
unfold(tensor, mode)
}
\arguments{
\item{tensor}{Three-mode tensor array.}

\item{mode}{Mode of operation.}
}
\value{
Matrix
}
\description{
Unfold/Matricize tensor. convert matrix to tensor by mode.
}
\examples{
X_i_jk = generate_dataset()$X_i_jk
X_i_j_k = fold(X_i_jk, mode=1, shape=c(I=8,J=5,K=4))
unfold(X_i_j_k, mode=1) # X_i_jk

}
